Remove ManhattanDepthLimited
the reason it was sometimes crashing at times is because it is simply not a valid/complete/admissible search (at least not the way I implemented it)
the manhattan value is optimistic and underestimates (obviously). this means that we might prune an existing valid branch because this node might be optimistically estimated better than the branch with a higher estimate, but lower actual cost

I don't see a way for this method to work 100% of the time since manhattan is not a true cost.